custom VFS: recover from dead semaphore owner (bb#1882)
when acquiring the SQLite lock succeeds after failing to acquire the semaphore,
it means that the owner of the semaphore is dead, or there was a race
condition. On a race condition the other process will release the
semaphore (it fails to acquire SQLite lock after succeeding to acquire
semaphore).
The current process owns the semaphore now. So make sure that it is properly released (waitsemHeld).

Reduce dead owner detection from 20s / DB-lock to ~10ms / DB-lock.